CROPSASSIT
CROPASSIT :A MOBILE APPLICATION FOR AGRICULTURISTS 

INTRODUCTION
“Agriculture is the backbone of India.” These were the words of laureate Jawaherlal Nehru the former prime minister of India. He wanted India to prosper in the agriculture sector as it was one of the major sectors in India. Today in this technologically advanced society the Agriculture sector has been struggling to keep up with the other sectors who are being benefited by the technology today's world has to offer. Mobile applications are one of  the technologies that can enhance farming in an unimaginable way. This app can guide agriculturists by predicting the weather, analysing the soil culture, intimating the underground water level, determining the highest market value for the crop in the region, and it offers descriptive videos through the process of farming. As per statistics of Government of India, over 12,000 suicides were reported in the agricultural sector every year since 2013, to curb this problem we introduced a feature in our app for interested people from all over the world to donate to the farmers who are in crisis, This also helps agriculturists to buy and rent different farming machinery for a sustainable price. The app allows to interact with other farmers on their experience in farming and the different farming techniques and tricks they used. People in the same region going in the same direction can club together and reduce the cost of transportation. This app helps to club these people together. The main objective of the app is to guide the farmers in every step of their journey of farming and contribute to the development of India.

INSTALLING THE APP
Download the .apk file and install it in your phone.

DESCRIPTION
CropAssit is an application oriented towards the enhancement of agriculture in developing countries like India. The app can be used by both agriculturists and citizens all around the world who are interested to donate towards the less privileged in the agricultural sector. For agriculturists the app provides a simple UI in two(currently) languages i.e Telugu and English for their better understanding. The app predicts the weather, analyses the soil culture, intimating the groundwater level by embedding Earth Enterprise, an open source release of Google Earth Enterprise, a geospatial application which provides the ability to build and host custom 3D globes and 2D maps. The app also has a feature to chat with other farmers on the platform to exchange their views and techniques in farming creating a sociable environment. Anyone interested to donate to the farmers can also access this app and chat with the farmers and know their problems personally and directly help the farmers eradicating all the middle men. In many ways this app can be used to be an assit in agriculture towards a united goal of enhancing the agriculture sector.